requirements.txt
streamlit
streamlit_chat
panda

